Product Backlog Item 8390791: Supporting INT and PPE environments for RP Task 8532262: Move INT api-version from prod to dogfood
This change:
Adds the 2020-08-20-beta API Version, which will be used for registering the ACS INT environment with RPaaS DF.
Adds x-ms-long-running-operation-options for ACS resource creation and deletion, which are async operations.
Removes x-ms-long-running-operation from EventGridFilter creation/deletion, which are synchronous.
Corrects the header in the delete.json examples to use Azure-AsyncOperation instead of Location.